Modified tests a bit. KM working except on M3 example
