--- src/test/resources/Answers.txt
@@ -1,109 +1,0 @@
-test1
-449
-(1,2)
-(2,4)
-(3,3)
-(4,1)
-(5,5)
-
-test2
-577
-(1,3)
-(2,4)
-(3,2)
-(4,6)
-(5,10)
-(6,7)
-(7,9)
-(8,1)
-(9,5)
-(10,8)
-
-test3
-17889
-(1,11)
-(2,20)
-(3,26)
-(4,12)
-(5,9)
-(6,25)
-(7,27)
-(8,23)
-(9,7)
-(10,14)
-(11,30)
-(12,3)
-(13,18)
-(14,17)
-(15,4)
-(16,19)
-(17,13)
-(18,1)
-(19,6)
-(20,15)
-(21,24)
-(22,21)
-(23,29)
-(24,22)
-(25,2)
-(26,5)
-(27,10)
-(28,8)
-(29,16)
-(30,28)
-
-test4
-473
-(1,37)
-(2,3)
-(3,48)
-(4,30)
-(5,47)
-(6,19)
-(7,32)
-(8,24)
-(9,9)
-(10,14)
-(11,26)
-(12,44)
-(13,50)
-(14,2)
-(15,11)
-(16,8)
-(17,7)
-(18,41)
-(19,45)
-(20,6)
-(21,28)
-(22,17)
-(23,1)
-(24,4)
-(25,18)
-(26,36)
-(27,40)
-(28,39)
-(29,20)
-(30,23)
-(31,16)
-(32,46)
-(33,43)
-(34,21)
-(35,27)
-(36,10)
-(37,15)
-(38,31)
-(39,49)
-(40,35)
-(41,38)
-(42,34)
-(43,5)
-(44,29)
-(45,22)
-(46,42)
-(47,33)
-(48,13)
-(49,25)
-(50,12)
-
-M3
-94
