+++ src/test/resources/answer1.txt
@@ -1,0 +1,6 @@
+449
+(1,2)
+(2,4)
+(3,3)
+(4,1)
+(5,5)
