+++ src/test/resources/answer2.txt
@@ -1,0 +1,11 @@
+577
+(1,3)
+(2,4)
+(3,2)
+(4,6)
+(5,10)
+(6,7)
+(7,9)
+(8,1)
+(9,5)
+(10,8)
