+++ src/test/resources/answer3.txt
@@ -1,0 +1,31 @@
+17889
+(1,11)
+(2,20)
+(3,26)
+(4,12)
+(5,9)
+(6,25)
+(7,27)
+(8,23)
+(9,7)
+(10,14)
+(11,30)
+(12,3)
+(13,18)
+(14,17)
+(15,4)
+(16,19)
+(17,13)
+(18,1)
+(19,6)
+(20,15)
+(21,24)
+(22,21)
+(23,29)
+(24,22)
+(25,2)
+(26,5)
+(27,10)
+(28,8)
+(29,16)
+(30,28)
