+++ src/test/resources/answer4.txt
@@ -1,0 +1,51 @@
+473
+(1,37)
+(2,3)
+(3,48)
+(4,30)
+(5,47)
+(6,19)
+(7,32)
+(8,24)
+(9,9)
+(10,14)
+(11,26)
+(12,44)
+(13,50)
+(14,2)
+(15,11)
+(16,8)
+(17,7)
+(18,41)
+(19,45)
+(20,6)
+(21,28)
+(22,17)
+(23,1)
+(24,4)
+(25,18)
+(26,36)
+(27,40)
+(28,39)
+(29,20)
+(30,23)
+(31,16)
+(32,46)
+(33,43)
+(34,21)
+(35,27)
+(36,10)
+(37,15)
+(38,31)
+(39,49)
+(40,35)
+(41,38)
+(42,34)
+(43,5)
+(44,29)
+(45,22)
+(46,42)
+(47,33)
+(48,13)
+(49,25)
+(50,12)
