+++ src/test/resources/answerM3.txt
@@ -1,0 +1,1 @@
+94
